requirements.txt
# YouTube Trailer Automation - Python Dependencies

# YouTube download
yt-dlp>=2024.1.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# RSS feed parsing (fallback for channel monitoring)
feedparser>=6.0.0
python-dateutil>=2.8.0

# FFmpeg wrapper (optional, we use subprocess directly)
# ffmpeg-python>=0.2.0
